Moves packages around to avoid clashses with the pre-existing dropwizard-health module
